docker/requirements.txt
gunicorn 
phonemizer
wave
torch
openai-whisper
scipy
numpy 
resampy
Levenshtein
librosa